companion/requirements.txt
websockets
keyboard
spotipy
python-dotenv
